0.7.0.1 codecleanup 转发群消息增加int? MessageId 基金助手增加尝试次数 色图下载响应使用CQ:reply 色图响应追加

+++ SheepQQBot3/CQExtensions.cs
@@ -11,7 +11,7 @@
         public static string Image(string filePath)
             => $"[CQ:image,file={filePath}]";
 
-        public static string Reply(long targetId, long messageId)
+        public static string Reply(long targetId, int messageId)
             => $"[CQ:reply,qq={targetId},id={messageId}]";
     }
 }
